+++ firebase-crashlytics-ndk/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [changed] Updated `firebase-crashlytics` dependency to 20.0.6
+
 # 20.0.5
 
 - [fixed] Fixed a runtime crash that could occur in minified native apps when using the Crashlytics